services/inference-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inference-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "confluent-kafka==2.5.3",
    "fastavro==1.9.7",
    "boto3==1.35.36",
    "tenacity==9.0.0",
    "prometheus-client==0.21.0",
    "structlog==24.4.0",
    "pydantic-settings==2.5.2",
    "numpy==1.26.4",
    "Pillow==10.4.0",
    "httpx==0.27.2",
]

[project.optional-dependencies]
ml = [
    "tritonclient[grpc]==2.49.0",
    "opencv-python-headless==4.10.0.84",
]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "httpx==0.27.2",
    "pytest-cov==5.0.0",
    "anyio==4.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/inference_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
